logistics-service/src/logistics/schemas/product.py
from pydantic import BaseModel, Field, ConfigDict
from decimal import Decimal
from typing import Optional
from datetime import datetime

# Base Schema: Shared properties
class ProductBase(BaseModel):
    name: str = Field(..., min_length=1, max_length=255)
    price: Decimal = Field(..., gt=0, decimal_places=2)
    stock_quantity: int = Field(..., ge=0)

class ProductCreate(ProductBase):
    pass

class ProductUpdate(BaseModel):
    name: Optional[str] = Field(None, min_length=1, max_length=255)
    price: Optional[Decimal] = Field(None, gt=0, decimal_places=2)
    stock_quantity: Optional[int] = Field(None, ge=0)

class ProductRead(ProductBase):
    id: int
    created_at: datetime
    updated_at: Optional[datetime] = None
    
    # We allow the model to read from ORM attributes
    model_config = ConfigDict(from_attributes=True)