logistics-service/pyproject.toml
[project]
name = "logistics-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.129.0",
    "psycopg>=3.3.2",
    "pydantic-settings>=2.12.0",
    "sqlalchemy[asyncio]>=2.0.46",
]
